Add a two-tier manifest for the public key crypto case -- also fix a bug
The idea is that the encryption (public) key can be "less secret" than
the decryption (private) key, but it is still useful -- and perhaps
essential -- to be able to get a list of which encrypted files are part
of a given manifest WITHOUT needing the decryption key.

So we'll include a baby manifest with just the encrypted filenames.