A dangerous script, not for the faint of heart
This is a first pass cleanup script for removing obsoleted backups, and
the blobs that are unique to them.

The idea is to compare the set of blobs per backup manifest, and discard
any old manifest whose similarity score is too high to a successive
backup. At the moment, this threshold is set at 90%, but this will
probably be a parameter. Also, the intent seems reasonable enough, but
the logic might be wrong -- there aren't tests yet.