At risk of being the dog who has no idea what he's doing at the computer
If we're going to support password entry in any capacity, using
something like argon2 (or perhaps: argon2) seems reasonable.

It's unclear to me how *stable* argon2 hashes are over time -- there's a
lot of params flying around in that standard encoding, and it's not
clear if I'm hooking into the perfect place for this. But I'm ok with
the tradeoff for now.